refactor: error imporvement of http-utils package and updated high category error format
re #1994
