+++ integration-tests/rpc-taquito.spec.ts
@@ -1,6 +1,6 @@
 import { CONFIGS } from "./config";
 import { RpcClient } from '@taquito/rpc';
-import { HttpResponseError } from "@taquito/http-utils";
+import { HttpResponseError } from "@taquito/core";
 
 
 CONFIGS().forEach(({ rpc, knownContract }) => {
