+++ packages/taquito-http-utils/package.json
@@ -60,7 +60,8 @@
   },
   "dependencies": {
     "@vespaiach/axios-fetch-adapter": "github:ecadlabs/axios-fetch-adapter",
-    "axios": "^0.26.0"
+    "axios": "^0.26.0",
+    "@taquito/core": "^15.0.1"
   },
   "devDependencies": {
     "@types/bluebird": "^3.5.36",
