+++ packages/taquito-remote-signer/package.json
@@ -63,6 +63,7 @@
   "dependencies": {
     "@stablelib/blake2b": "^1.0.1",
     "@stablelib/ed25519": "^1.0.3",
+    "@taquito/core": "^15.0.1",
     "@taquito/http-utils": "^15.0.1",
     "@taquito/taquito": "^15.0.1",
     "@taquito/utils": "^15.0.1",
